Switch from @1 numbered params to @ single implicit param
By far the most common use case for implicit parameters is for a
single parameter.  Using multiple implicit params instead of
explicitly named params is more likely to lead to code that is
harder to understand, since the argument order must be implicitly
understood.

This change makes the single argument case simpler (@1 -> @), and
avoids the problems that occur when multiple arguments are used,
by requiring explicit arguments in that case.