Fix {**{}} and [**{}] compilation
{**{}} wasn't checking whether we were compiling a hash or
using keyword arguments, and resulted in an empty frozen
hash instead of an empty unfrozen hash.  Check whether we
are compiling a hash or using keywords and only apply the
empty keyword optimization if we are compiling keywords.

Unfortunately, [**{}] parses similar to keywords instead of
as a literal hash.  Change that so it parses as a literal
hash, treating it the same as [{**{}}].  This changes the
behavior so that empty keyword splats in array literals
are no longer removed.  That is inconsistent with the
behavior of empty keywords in method calls, but it is more
compatible with Ruby 2.6 and previous versions.

The newarraykwsplat VM instruction can probably be removed
after this commit, since keyword splats in arrays are now treated
as literal hashes instead of like method keywords.