Modify copy coroutine to avoid crashes
This fixes fiber usage on OpenBSD/sparc64 (which uses the copy coroutine).

There are three related changes:

1) Don't flush register windows when saving the stack until right before
copying the stack to the heap. This fixes local variables not being
correct after the stack is restored. This idea was taken directly from
the Ruby 2.6 implementation.

2) Add a char[128] buffer to the coroutine_save_stack function, similar
to the VALUE[STACK_PAD_SIZE] buffer in Ruby 2.6.

3) Make sure we copy the entire current stack frame before copying the
stack to the heap by calling a function and returning the address of the
local variable inside of it (this causes a compiler warning). This is
necessary to ensure we don't have a partial copy of the current stack
frame. Ruby 2.6 does something similar using a recursive function call.

4) Smear volatile over the pointers so this works without disabling
optimization.  The Ruby 2.6 implementation also used volatile.

Fixes [Bug #16809]