Do not restart enumerator if it prematurely terminates due to exception
Previously, if an exception occurred during enumeration, the next
call to Enumerator#next and similar methods would restart the
Enumerator from the beginning, which is unlikely to be the desired
behavior.

This changes the behavior so that after the enumerator has raised
an exception, future calls to Enumerator#next and similar methods
will raise StopIteration, similar to when calling Enumerator#next
after the enumerator has finished.

Fixes [Bug #16816]