Warn when keyword label is followed by a newline
The behavior will change in Ruby 3.3 to treat the call as an implicit
keyword.  The following code:

```ruby
x = 1
p x:
2
```

Currently prints `{x: 2}`, but in Ruby 3.3, will print `{x: 1}`. This
changes the parser to print a warning in such cases, allowing people
to fix their code in Ruby 3.2 before the change in Ruby 3.3.

Fixes [Bug #18396]

Co-authored-by: Yusuke Endoh <mame@ruby-lang.org>