Eliminate array allocation for f(arg, *a)
Previously, the splatarray instruction had a true argument in the
case (duplicating the array), but as the array is not being
modified, there is no reason to duplicate it.

This changes the dup_rest argument of setup_args_core to be a
pointer, so that there is a single place in memory where we can
track whether we have duplicated an array when splatting. Once
an array has been duplicated when splatting, the pointer is
updated so that future arrays will not be duplicated when
splatting.

The previous approach never had multiple splatarray instructions
with true arguments, but the design made it impossible to tell
the difference between:

  f(arg, *a)     # No array allocation needed
  f(arg, *a, *b) # Array allocation needed

Note that you cannot currently eliminate the allocation with:

  f(arg, *a, &block)

As the block expression could modify splatted argument.