Avoid array allocation for f(*a, **kw)
This avoids array allocation (duplicating `a` via `splatarray true`)
for the cases where `kw` is a local variable and a hash.

In other cases, you cannot safely avoid the allocation, because
evaluation requires the following order:

* `a`
* `a.to_a`/`a.dup` # to_a if a is not array, and dup if necessary
* `kw`
* `kw.to_hash` # if kw is not hash

If `kw` is an expression, method call, or constant reference, it
may change the result of `a.to_a`. If `kw` is not a hash, calling
`kw.to_hash` may change the result of `a.to_a`. If `a` is not an
array, calling `a.to_a` may change the value of `kw` if `kw`
is not a local variable.

This is implemented via a new splatarrayiftype instruction. This
instruction checks the type of the top argument on the stack,
and compares it to the argument.  If they match, it operates
like `splatarray false` (not duplicating the splatted array).
If they do not match, it operates like `splatarray true`
(duplicating the splatted array).

For `a = kw = nil; f(*a, **kw)`:

Instructions Before:

```
== disasm: #<ISeq:<main>@-e:1 (1,0)-(1,25)>
local table (size: 2, argc: 0 [opts: 0, rest: -1, post: 0, block: -1, kw: -1@-1, kwrest: -1])
[ 2] a@0        [ 1] kw@1
0000 putnil                                                           (   1)[Li]
0001 dup
0002 setlocal_WC_0                          kw@1
0004 setlocal_WC_0                          a@0
0006 putself
0007 getlocal_WC_0                          a@0
0009 splatarray                             true
0011 getlocal_WC_0                          kw@1
0013 opt_send_without_block                 <calldata!mid:f, argc:2, ARGS_SPLAT|FCALL|KW_SPLAT>
0015 leave
```

Instructions After:

```
== disasm: #<ISeq:<main>@-e:1 (1,0)-(1,25)>
local table (size: 2, argc: 0 [opts: 0, rest: -1, post: 0, block: -1, kw: -1@-1, kwrest: -1])
[ 2] a@0        [ 1] kw@1
0000 putnil                                                           (   1)[Li]
0001 dup
0002 setlocal_WC_0                          kw@1
0004 setlocal_WC_0                          a@0
0006 putself
0007 getlocal_WC_0                          a@0
0009 getlocal_WC_0                          kw@1
0011 splatarrayiftype                       8
0013 opt_send_without_block                 <calldata!mid:f, argc:2, ARGS_SPLAT|FCALL|KW_SPLAT>
0015 leave
```

There is one pathological case where this optimization
results in different behavior. That is when `a` is not
an array, and `a.to_a` modifies the local variable
value via a binding:

```ruby
class C
  def initialize(binding)
    @binding = binding
  end

  def to_a
    @binding.local_variable_set(:d, {b: 1})
    []
  end

  def f(*x, **a)
    a
  end
end

d = {a: 2}
c = C.new(binding)
p c.f(*c, **d)
```

This is because the optimization technically changes the
evaluation order to:

* `a`
* `kw`
* `a.to_a` # if a is not array
* `kw.to_hash` # if kw is not hash

However, if `kw` is a local variable, evaluating it will never
change `a.to_a`, and `a.to_a` can only change the value of `kw`
if it is a local variable by modifying the related binding.
Since it seems incredibly unlikely anyone would be relying on
such behavior, and considering that `f(*a, **kw)` where `kw`
is a local variable is a very common pattern in Ruby code,
I think the benefits of the optimization far exceeds the cost
of breaking this pathological case.