Release GVL around pw and gr function calls in Etc
These can block as they parse files.

I'm not sure if other function calls in Etc can block, so this
doesn't add GVL releasing to them.

Releasing the GVL for these methods results in thread-safety
issues, since thread-safety for the methods relied previously
on the GVL.  Add a mutex and surround GVL-releasing calls
with a mutex.