Avoid array allocation in 2 types of method calls
This implements 2 separate but related optimizations. The first
optimization is caller-side, when splatting nil:

```ruby
a(*nil)
```

This previously allocated an array, but I have determined this
array allocation is unnecessary.  The instructions in this case are:

```
0000 putself                                                          (   1)[Li]
0001 putnil
0002 splatarray                             false
0004 opt_send_without_block                 <calldata!mid:a, argc:1, ARGS_SPLAT|FCALL>
0006 leave
```

The method call uses `ARGS_SPLAT` without `ARGS_SPLAT_MUT`, so the
returned array doesn't need to be mutable.  I believe all cases where
`splatarray false` are used allow the returned object to be frozen,
since the `false` means to not duplicate the array.  The optimization
in this case is to have `splatarray false` push a shared empty frozen
array, instead of calling `nil.to_a` to return a newly allocated array.

There is a slightly backwards incompatibility with this optimization,
in that `nil.to_a` is not called.  However, I believe the new behavior
of `*nil` not calling `nil.to_a` is more consistent with how `**nil`
does not call `nil.to_hash`.  Also, so much Ruby code would break if
`nil.to_a` returned something different from the empty hash, that it's
difficult to imagine anyone actually doing that.

The second optimization is for methods with anonymous positional splat
parameters, when called without arguments:

```ruby
def a(*) end
a()
```

This used to allocate an array for the anonymous positional splat.
The optimization works similarly to the `*nil` optimization, using
a shared empty frozen array instead in this case.  Because the
anonymous positional splat can never be referenced, using a shared
frozen array in this case works.  This is similar to how anonymous
keyword splats work, except those use `nil` instead of a shared
empty frozen hash.

This updates the allocation tests to check that the array allocations
are avoided where possible.

The only breakage in `make test-all` is in a Ractor test, which was
relying on an anonymous positional splat parameter being mutable when
called with no arguments (and thus, raising a Ractor::IsolationError).
I updated the test to call the method with a splat of a non-frozen
array, so it still tests the same thing.